Fixes in RNA-seq-analysis vignette and cluster_hits() function
In the RNA-seq-analysis vignette, part of the code block of the cluster_hits() function was wrong. It is not executed for the vignette, therefore, it was not evident before.
The cluster_hits() function contained a bug, were the double spline plots were not generated in the report when the condition column was a factor and not a character. This was because it compared the unique values of the condition column with the names of dataframes, and when the condition column is a factor, it finds no match. Now, it converts it to character before doing those things.