extract_data() fun can now handle NA values
Before, it was not able to identify the 6x6 "starting seed" correctly, because it searched for a 6x6 field without NA values. Now it accepts NA values to a high degree (it does not accept only NA  values, because empty parts in an Excel document for example are filled with NA values, so that would cause issues.