Added detect_excursions() exported function
For now, the file detect_excursions.R contains the detect_excursion() function and the plot_excursions() function.
The first deals with finding excursions (local outlashes up or down) with two pairwise tests of timepoints.
Only when for example T1 vs. T2 is significant with limma and T2 vs. T3 also, but in a different direction, it is a valid excusrion. So only if it goes either up and then down or down and then up, and both times significant, it is detected.
This is to complement the splines, which are weak at detecting such patterns. The plotting function then just creates scatter plots showing the excusrions in red (the corresponding timepoints) and the rest in grey