find_peaks_valleys()
Made new file in the source code R folder (find_peaks_valleys.R), that contains the function find_peaks_valleys() that is available to the user.
This function then allows to receive an HTML report where all the significant peaks and valleys (one timepoint deviations) are shown. For now, the HTML report logic is not yet in there, but it does the UI test, and the plotting.