Robust fit
When the linear model assumption of homoscedasticity is violated, it can invalidate the resulting adj.pvalues.
Therefore, now the user can decide to explicitly use the robust fit strategy or let this be handled implicitly in the code.
There is a strategy for RNA-seq data, which just involves running voomWithQualityWeights instead (voomWithDreamWeights apparently handles this anyways).
Also, there is a strategy for non-RNA-seq data, which involves using  arrayWeights() and eBayes(robust = TRUE).
The implicit fall back logic uses a Wilcox test to check if the samples in both levels have different variance.