Refactored between level code run_limma_splines()
This code created all level pairs of the condition factor, and then always took the respective part of the data, fitted a model and extracted the results from there. However, you cannot split the voom object up, and this only worked for the voom object because I never had more than two levels in the condition.
Now, it just always fits a global model with all the data for the between level analysis, and then afterwards loops over all the comparisons and gets the results. This now works for both non-RNA-seq data as well as for RNA-seq data. I confirmed that this new code produces the exact same output as the code from before.