Refactored find_peaks_valleys()
There was a bug in there, that skipped a hit when there was an NA value in the timepoint of the hit or one of the neighboring timepoints.
Now, it does not filter the hits based on that both left and right must be both higher or lower, but it classifies them into peak, valley, top of cliff, bottom of cliff based on a comparison logic.
Split up the code into smaller functions so that the peak_valley_test() function has a single responsibility.
Also, renamed some variables.

In addition, just made a very small change to the preprocess_rna_seq_data, because this did still contain the old implementation of the SplineOmics::check_homoscedasticity_violation() function.