find_pvc() HTML report
find_pvc() now generates an HTML report. It has to still be refined a bit, but otherwise, I am almost there.

Additional small things:
Fixed a bug in the cluster_hits() function that warned the user about a large report that will be generated, even though the argument report was set to FALSE.
All the variables in the former run_gsea() function were renamed from gsea to ora (overrepresentation analysis) which is the correct term for clusterProfiler.