Updated check_homoscedasticity_violation()
Now, the function uses Levenes test instead of the Breusch-Pagan test.
The big difference is that the function now takes the residuals of the actual spline models with the experimental design specified by the user.
Those residuals are split into the groups, that are defined by the experimental condition (for example phase exponential and stationary (two levels within condition).
Levenes test is then used to check if the variance between the residuals within a group are significantly different between the groups.