Refactored and improved compare_results() (former correlate_results) fun
This function now also makes a correlation plot just for the hits (significant features).
Further, it compares the hits by counting the overlap, and calculating the Jaccard-Index for the overlap.
Additionally, the whole function was refactored: Code was moved to modular functions (like the plotting code) and now also the inputs are controlled.
Also, now the user can specify names for the SplineOmics objects, that are then shown in the axis descriptions in the plots. Also, adj.p-value thresholds can be provided (per default 0.05) which allow to do all the hit comparisons.