More info in run_ora() reports
The HTML reports generated by the run_ora() function contained too little information about the inputs and the results to which the ora report is connected.
To adress this, the ora report now contains downloadable files with the foreground and background genes, reports the clusterProfiler parameters used, and the run_ora() function has a new argument which allows to pass a string which is the name of the corresponding cluster_hits() HTML report, that contains the results that are the basis for the respective run_ora() report.

In addition, two small bugs were fixed in the unit tests for the cluster_hits() and the run_limma_splines() functions. Further, the run_ora() code again was reset to an older version. I think this is because RStudio loads an older state. This behavior in RStudio was now turned off, and the newer code from GitHub again used to replace the code that was resetted.