Auto-dof with loocv
Introduced new capability, that automatically finds the best spline dof based on the error on loocv. For every linear model fit, it performs the analytical loocv and then looks at the error sum of all features. It does this with every dof from 2 to min(timepoints, 10). It is mostly done, but to fully finish it so that it also returns the selected dof properly, something else needs to be refactored, which is the separation of the computations for the isolated and integrated approach. Therefore, I now commit this, then refactor this other logic, commit, and the followup commit will finish the auto-dof.