Extendet unit tests of run_limma_splines()
Before, there was just one test, and it just checked the structure of the SplineOmics object.
Now, there are two tests for run_limma_splines(): One for the isolated, and one for the integrated mode.
Both now don't just check the structure, but also the amount of hits in all results, as well as 10 randomly chosen values in the first 6 columns of all result tibbles. Those 10 values must match with the hardcoded values in the test.