Fixed new integrated-isolated separated approach for random effects
The problem was that the new function extract_within_level_time_effects() was not adjusted for the random effects, that are handeled with the varianePartition package functions. Now it uses them and it works out. And I fixed another issue where the clustering report needed the fit object from limma (or variancePartition) to make the spline plots properly for the RNAseq results. One issue still remaining is that the spline plots look off for several glycan features. I still have to fix this.