Introduced auto-dof into preprocess_rna_seq_data()
This functionality was only introduced in limma, but not in the rna preprocesssing function. Therefore, the auto-dof functionality did not work when using RNA-seq data. This is now solved. The preprocessing function now also tries out the optimal dof with LOOCV. Because of this, now not just the voom_obj is returned, but it is directly inserted into the splineomics object, which is optinallly updated with the optimal dof, and that is returned.