Made cluster_hits() usable for many hits
Introduced a new function argument, called max_hit_number. This controls the plotting logic of the report generation, so it limits the amount of plots and amount of detail shown in the plots to a maximum number of hits per cluster. This allows to also generate clustering reports for analyses with many hits, because nowbody anyways looks at thousands of spline plots.