Fixed issue with spline plots in cluster_hits() report HTML
The problem was that the intercept looked really off for many spline plots of time effect. This was because variancePartition::dream() did not run successfully for all features, which then causes a mismatch between data and the coefficients. The coefficients before had just a increasing number as descriptions (so I mean the coefficient rows of beta) and this then caused the misalignment. Further, there was another bug that did not show the plots of the average diff condition and interaction condition time. This was because they needed looping over the hits of time_effect, and the max_hit_number regulated that, instead of regulating how many plots were generated. This was also fixed.