Checking rowheaders of data and removing batch effect correction for plotting
Added check that ensures that all data matrix rowheaders are unique among the rowheaders.

Removed the function remove_batch_effect_cluster_hits() from cluster_hits(). This was done because the code no longer did the batch effect correction for the plotting data. Rather, the intercept of the predicted spline curves is adjusted with a least squares fit of the spline shape to the corresponding data row (does not change the splines, just corrects the intercept for visual purposes only. The intercept is not subject to any statistical tests or clustering.