Introduced clusterProfiler::enrichGO() into SplineOmics::run_ora()
Now, there is a new argument of run_ora(), namely enrichGO_cfg. This allowes to specify which databases passed are from BioConductor GO, along with the arguments OrgDb, keyType, and ontology. For those databases, inside run_ora(), conditionally the enrichGO function is run, and afterwards the simplify function (to collapse hierarchical terms).