Added new unit test, and fixed bug in run_limma_splines() and cluster_hits() when the design interaction terms were flipped in order.
The finding of the correct design columns in several parts of the code assumed that condition always comes before time. This is of course fragile, because this ordering is arbitrarily defined by the user and there is no correct order. So far, it was just tested with one ordering, and that worked out, but it failed for other orders. Now, this is fixed.

This was discovered by adding a new unit test, for the cluster_hits() function wiht the limma integrated mode (a design with an interaction term).