Refactored and extended Levene's test for heteroscedasticity
Now, the Levene's test is not only performed feature wise, with the condition as the group, but also sample wise. This is because the array weights from limma actually work sample wise, so it makes sense to test heteroscedasticity there. Additionally, due to these changes, the respective Levene test function was refactored.