use_array_weights for mode = isolated
Now also for the mode = isolated it is possible to use array weights. For this, it was necessary to include the array weights logic into the core limma results generation code for the isolated mode, and enable the processing of the info of the levene's test until the report generation.