Reverted png heatmap switch in plot2base64()
This change was introduced, because on Mac, there is a problem with rendering the ComplexHeatmap plot in the R CMD environment (without graphic device). I tried to solve this plot by switching to png image from .sgv, but it did not work. That is why I am reverting this change now.