Tried again to fix Mac ComplexHeatmap plot issue, and added new unit tests
Now I use ComplexHeatmap::draw for the plot, instead of simply print. This is done conditionally for the ComplexHeatmap plots.

Further, added two new unit tests, that are equivalent to the existing unit tests for the cluster hits function for the integrated and isolated approach, but have use_array_weights = NULL, for the implicit logic (to test that).