Changed returns of cluster_hits()
Before, the returns of the cluster_hits() function were messy. Now, it returns just one main result table, and the plots.
Note that in this commit there are already changes of the run_ora() function, because I forgot to commit once I finished the changes for the cluster_hits() function.