Added min_effect_size to cluster_hits()
Now, one can also specifiy an effect size threshold for the cluster_hits() function. This is then used to filter the limma results in addition to the significance cutoff. This was done because with enough samples, also tiny, irrelevant differences can be detected. The changes in the code only deal with this new addition, and it is basically done. The code just has to be reviewed a little more, but the thresholding is in place and the cumulative travel and the avg diff conditions are shown in the HTML reports already