Double spline clustering and cat3 effect size
Improved the code of the effect size for the splines. Additionally, added the functionality of the double spline clustering for category 3 (cat3) that allows to assign clusters to the hits in that category for the following enrichment. Further, changed the effect size metric for cat3. Now, the thresholding is done for the differential cumulative travel.