Cluster quality metrics
Each spline of category 1 (time effect) is now compared to the respective cluster centroid (average spline) via Spearman correlation. This is delivering a "goodness-of-fit" metric for each spline, and also a quality metric for the respective cluster (average of all the Spearman correlations to the cluster centroids of the splines in that cluster). The individual correlations are shown in the time effect spline plots, and the distribution of the corrs within a cluster is shown as a density plot in the report.