z-score norm splines & new cluster quality metric
Changed to normalize the curves with z-score across the time before the clustering, because this better captures also the size of change, and not just qualitatively the changes. Furher, the new cluster quality metric is the silhouette score, changed the plotting and the computation of the quality scores for that.