Improved documentation and adressed R CMD check errors
For the cluster_hits() function, the documentation was a bit improved. Further, there were a set of issues for the R CMD check, also specific to the GitHub actions workflow R CMD check. I tried to adress all of them and to also clean things up that are part of that (such as removing dependencies that are no longer used).