Added download_bioconductor_database() to exported functions
The function was there before, but I forgot to add it to the user facing functions. Now, I improved a bit the code of it and the documentation, and also slightly refined the code of the download_enrichr_databases() function and the run_ora() function, and refined the get_started.rmd vignette a bit.