Fixed two bugs and added effect sizes to topTables
In the extract_data() function, there was problematic code line that removed NA values from the feature names vector, which however then led to a mismatch with the data, which would stop the function with an error.
In explore_data, there also was a problem at one point for the autocorrelation computation, that stopped with an uncaught error.
In cluster_hits(), now the effect sizes are added to the topTables