Fixed plot_info
Before, there was a bug with the values of plot_info$treatment_labels and plot_info$treatment_timepoints. This has now been resolved. The input control was updated, and now one has to explicitly specifiy for which of the two time effect levels one wants to have the treatment lines, and also if one wants to have the lines for the double spline plots.