BiocCheck criteria
Worked on fulfilling the BiocCheck criteria. For that, introduced examples in several functions. Based on that, discovered suboptimal code or bugs, fixed them. One major thing, dof = 1 had no support, because it has another column name (X and not X1) and requires a t-test and not an F-test. Fixed that too.