Fixed bug in cluster_hits() and changed batch effect correction in find_pvc()
In the cluster_hits() function, there was a bug that prevented the cT values to be written into the hits table. This was because the code had a hardcoded meta column name for the condition and did not work when the name was different.
The batch effect correction in the find_pvc() function was changed to use the limma::removeBatchEffect() function again, instead of writing the batch effect in the design. This is because for the PVC-test, only this prevents many artifacts in the presence of a batch effect.