Changed adj_pthresholds and plot_info arguments of cluster_hits()
adj_pthresholds was renamed to adj_pthresh_time_effect and now just supports a single threshold for all conditions. plot_info does not accept the double_spline_plots value in the treatment_labels and treatment_timepoints lists. Instead, it generates the treatment lines for the double spline plots based on the treatment lines for the time effect.