Fixed and refactored run_limma_splines() contrasts for mode = integrated
There was a bug for the multi-condition scenario for the mode = integrated in the function run_limma_splines(). The pairwise contrasts for result category 2 and 3 were not created correctly. This was now fixed for the limma part only (variancePartition part is coming next). For this, the contrast procedures were also refactored. Now, a joint function creates all contrasts before the fitting function (because variancePartition needs the contrast matrix already at fitting time). Now the logic is cleaner and unified.